+++ src/main/java/org/dependencytrack/model/vuln_vb/Product.java
@@ -1,4 +1,4 @@
-package org.dependencytrack.model.vulnDb;
+package org.dependencytrack.model.vuln_vb;
 
 
 import java.util.ArrayList;
