+++ src/main/java/org/dependencytrack/parser/vulndb/ModelConverter.java
@@ -21,16 +21,16 @@ package org.dependencytrack.parser.vulndb;
 import alpine.common.logging.Logger;
 import org.apache.commons.lang3.StringUtils;
 import org.dependencytrack.model.Cwe;
-import org.dependencytrack.model.vulnDb.Author;
-import org.dependencytrack.model.vulnDb.ExternalReference;
+import org.dependencytrack.model.vuln_vb.Author;
+import org.dependencytrack.model.vuln_vb.ExternalReference;
 import org.dependencytrack.model.Vulnerability;
 import org.dependencytrack.parser.common.resolver.CweResolver;
 import org.dependencytrack.persistence.QueryManager;
 import us.springett.cvss.CvssV2;
 import us.springett.cvss.CvssV3;
 import us.springett.cvss.Score;
-import org.dependencytrack.model.vulnDb.CvssV2Metric;
-import org.dependencytrack.model.vulnDb.CvssV3Metric;
+import org.dependencytrack.model.vuln_vb.CvssV2Metric;
+import org.dependencytrack.model.vuln_vb.CvssV3Metric;
 
 import java.math.BigDecimal;
 import java.time.OffsetDateTime;
@@ -56,7 +56,7 @@ public final class ModelConverter {
      * @param vulnDbVuln the VulnDB vulnerability to convert
      * @return a Dependency-Track Vulnerability object
      */
-    public static Vulnerability convert(final QueryManager qm, final org.dependencytrack.model.vulnDb.Vulnerability vulnDbVuln) {
+    public static Vulnerability convert(final QueryManager qm, final org.dependencytrack.model.vuln_vb.Vulnerability vulnDbVuln) {
         final org.dependencytrack.model.Vulnerability vuln = new org.dependencytrack.model.Vulnerability();
         vuln.setSource(org.dependencytrack.model.Vulnerability.Source.VULNDB);
         vuln.setVulnId(sanitize(String.valueOf(vulnDbVuln.getId())));
