+++ src/main/java/org/dependencytrack/util/VulnDBUtil.java
@@ -17,20 +17,20 @@ import org.apache.http.util.EntityUtils;
 import org.dependencytrack.common.HttpClientPool;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.dependencytrack.model.vulnDb.Cpe;
-import org.dependencytrack.model.vulnDb.Product;
-import org.dependencytrack.model.vulnDb.Results;
-import org.dependencytrack.model.vulnDb.Vendor;
-import org.dependencytrack.model.vulnDb.Version;
-import org.dependencytrack.model.vulnDb.ApiObject;
-import org.dependencytrack.model.vulnDb.Vulnerability;
-import org.dependencytrack.model.vulnDb.Classification;
-import org.dependencytrack.model.vulnDb.Author;
-import org.dependencytrack.model.vulnDb.CvssV2Metric;
-import org.dependencytrack.model.vulnDb.CvssV3Metric;
-import org.dependencytrack.model.vulnDb.NvdAdditionalInfo;
-import org.dependencytrack.model.vulnDb.ExternalText;
-import org.dependencytrack.model.vulnDb.ExternalReference;
+import org.dependencytrack.model.vuln_vb.Cpe;
+import org.dependencytrack.model.vuln_vb.Product;
+import org.dependencytrack.model.vuln_vb.Results;
+import org.dependencytrack.model.vuln_vb.Vendor;
+import org.dependencytrack.model.vuln_vb.Version;
+import org.dependencytrack.model.vuln_vb.ApiObject;
+import org.dependencytrack.model.vuln_vb.Vulnerability;
+import org.dependencytrack.model.vuln_vb.Classification;
+import org.dependencytrack.model.vuln_vb.Author;
+import org.dependencytrack.model.vuln_vb.CvssV2Metric;
+import org.dependencytrack.model.vuln_vb.CvssV3Metric;
+import org.dependencytrack.model.vuln_vb.NvdAdditionalInfo;
+import org.dependencytrack.model.vuln_vb.ExternalText;
+import org.dependencytrack.model.vuln_vb.ExternalReference;
 
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
@@ -39,7 +39,6 @@ import java.net.URISyntaxException;
 import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
 public class VulnDBUtil {
